requirements.txt
autogen-agentchat
autogen-ext[tools]
mcp
python-dotenv
openai
fastapi[standard]==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
